[fix] 로그아웃 시 Cookie 삭제 안되는 버그 수정

+++ board-back/src/main/java/com/solpooh/boardback/entity/primaryKey/FavoritePk.java
@@ -17,7 +17,7 @@ public class FavoritePk implements Serializable {
     @Column(name = "user_email")
     private String userEmail;
     @Column(name = "board_number")
-    private int boardNumber;
+    private Long boardNumber;
 
     @Override
     public boolean equals(Object o) {
